Fix guess matching bug
Fixes a bug when checking if a string agrees with a particular guess,
where if the guess coantained a duplicate letter where one of them was
an exact match and the other was not in the word (i.e. one green and
one gray), then all strings compared against it that contained that
letter would be considered not matching.

We now correctly ignore any letters in the string that have already
been "used" as exact matches to avoid trying to match a letter instance
as exact and nonexistent at the same time.